requirements.txt
# --- Configuration & Environment ---
hydra-core
python-dotenv
omegaconf

# --- Experiment Tracking & Logging ---
wandb
tqdm

# --- Data Processing ---
numpy
pandas
scikit-learn
pymongo

# --- Optional but Recommended ---
matplotlib
pillow